visualifyai-ai/requirements/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
openai==1.82.0
sqlalchemy==2.0.41
asyncpg==0.30.0
pdfplumber==0.11.6
python-multipart==0.0.20
transformers==4.36.0
python-docx==1.1.2
torch==2.5.1